fixed tests, further organization
